+++ traffic_ops/ort/atstccfg/cfgfile/remapdotconfig.go
@@ -201,6 +201,7 @@ func GetConfigFileServerRemapDotConfig(cfg config.TCCfg, serverNameOrID string) 
 				Protocol:                 ds.Protocol,
 				AnonymousBlockingEnabled: ds.AnonymousBlockingEnabled,
 				Active:                   *ds.Active,
+				RangeSliceBlockSize:     ds.RangeSliceBlockSize,
 			})
 		}
 	}